(chores) camel-core: fix non-atomic operations on volatile fields

+++ core/camel-support/src/test/java/org/apache/camel/support/task/ForegroundTaskTest.java
@@ -45,7 +45,7 @@ public class ForegroundTaskTest extends TaskTestSupport {
                 .build();
 
         task.run(this::booleanSupplier);
-        assertEquals(maxIterations, taskCount);
+        assertEquals(maxIterations, taskCount.intValue());
 
         Duration duration = task.elapsed();
         assertNotNull(duration);
@@ -67,7 +67,7 @@ public class ForegroundTaskTest extends TaskTestSupport {
                 .build();
 
         task.run(this::booleanSupplier);
-        assertEquals(maxIterations, taskCount);
+        assertEquals(maxIterations, taskCount.intValue());
     }
 
     @DisplayName("Test that the task does not run for more than the max iterations when using a predicate and an initial delay")
@@ -84,7 +84,7 @@ public class ForegroundTaskTest extends TaskTestSupport {
                 .build();
 
         task.run(this::taskPredicate, new Object());
-        assertEquals(maxIterations, taskCount);
+        assertEquals(maxIterations, taskCount.intValue());
     }
 
     @DisplayName("Test that the task stops running once the predicate is true")
@@ -101,7 +101,7 @@ public class ForegroundTaskTest extends TaskTestSupport {
                 .build();
 
         task.run(this::taskPredicateWithDeterministicStop, Integer.valueOf(3));
-        assertEquals(3, taskCount);
+        assertEquals(3, taskCount.intValue());
     }
 
     @DisplayName("Test that the task stops running once the predicate is true when the test is slow")
@@ -118,7 +118,7 @@ public class ForegroundTaskTest extends TaskTestSupport {
                 .build();
 
         task.run(this::taskPredicateWithDeterministicStopSlow, Integer.valueOf(3));
-        assertTrue(taskCount < maxIterations);
+        assertTrue(taskCount.intValue() < maxIterations);
     }
 
     @DisplayName("Test that the task stops running once the predicate is true when the test is slow")
@@ -135,6 +135,6 @@ public class ForegroundTaskTest extends TaskTestSupport {
                 .build();
 
         task.run(this::taskPredicateWithDeterministicStopSlow, Integer.valueOf(3));
-        assertTrue(taskCount < maxIterations);
+        assertTrue(taskCount.intValue() < maxIterations);
     }
 }
